+++ src/loaders/loaders/fonts.js
@@ -1,0 +1,6 @@
+export default function (options) {
+    return {
+        test: /\.(ttf|eot)/,
+        loader: 'url',
+    };
+};