+++ src/loaders/loaders/html.js
@@ -1,0 +1,6 @@
+export default function (options) {
+    return {
+        test: /\.html$/,
+        loader: options.angular ? 'ngtemplate!html?removeRedundantAttributes=false' : 'html',
+    };
+};