+++ src/loaders/loaders/json.js
@@ -1,0 +1,6 @@
+export default function (options) {
+    return {
+        test: /\.json/,
+        loader: 'json',
+    };
+};