+++ src/loaders/loaders/ts.js
@@ -1,0 +1,6 @@
+export default function (options) {
+    return {
+        test: /\.ts$/,
+        loader: options.angular ? 'ng-annotate!' + options.loaders.ts : options.loaders.ts,
+    };
+};