+++ src/loaders/loaders/webfonts.js
@@ -1,0 +1,10 @@
+export default function (options) {
+    return {
+        test: /\.woff/,
+        loader: 'url',
+        query: {
+            limit: options.inlineLimit,
+            mimetype: 'application/font-woff',
+        },
+    };
+};