+++ src/loaders/preloaders/eslint.js
@@ -1,0 +1,7 @@
+export default function (options) {
+    return {
+        test: /\.js$/,
+        loader: "eslint-loader",
+        exclude: /node_modules/,
+    };
+};