+++ src/plugins/occurenceOrder.js
@@ -1,0 +1,5 @@
+import webpack from 'webpack';
+
+export default function (options) {
+    return new webpack.optimize.OccurenceOrderPlugin(true);
+}