+++ src/plugins/provide.js
@@ -1,0 +1,9 @@
+import webpack from 'webpack';
+
+export default function (options) {
+    return new webpack.ProvidePlugin({
+        $: 'jquery',
+        _: 'lodash',
+        jQuery: 'jquery',
+    });
+};