+++ src/plugins/stats.js
@@ -1,0 +1,7 @@
+import StatsPlugin from 'stats-webpack-plugin';
+
+export default function (options) {
+    return new StatsPlugin('stats.json', {
+        chunkModules: true,
+    });
+};