+++ src/templates/packages.js
@@ -1,0 +1,20 @@
+export default function (config, options, loaders, plugins) {
+    return config.merge({
+        module: {
+            loaders: [
+                loaders.js,
+                loaders.json,
+            ],
+        },
+        output: {
+            path: options.outputPath,
+            filename: 'index.js',
+            library: options.libraryName,
+            libraryTarget: 'umd'
+        },
+        plugins: [
+            plugins.occurenceOrder,
+            plugins.define,
+        ]
+    });
+};